more size_t conversions, albeit relatively unused fields

+++ include/rt/rt_instance.h
@@ -107,8 +107,8 @@ struct rt_i {
     size_t              rti_cut_totobj; /**< @brief  # objs in all bins, total */
     size_t              rti_cut_maxdepth;/**< @brief  max depth of cut tree */
     struct soltab **    rti_sol_by_type[ID_MAX_SOLID+1];
-    int                 rti_nsol_by_type[ID_MAX_SOLID+1];
-    int                 rti_maxsol_by_type;
+    size_t              rti_nsol_by_type[ID_MAX_SOLID+1];
+    size_t              rti_maxsol_by_type;
     int                 rti_air_discards;/**< @brief  # of air regions discarded */
     struct bu_hist      rti_hist_cellsize; /**< @brief  occupancy of cut cells */
     struct bu_hist      rti_hist_cell_pieces; /**< @brief  solid pieces per cell */
